LayoutTests/http/tests/media/fairplay/legacy-fairplay-hls-expected.txt


EVENT(webkitneedkey)
FETCH: resources/cert.der OK
EVENT(webkitkeymessage)
PROMISE: licenseResponse resolved
EVENT(canplaythrough)
END OF TEST
